Include jcip annotations ref'd by commons http, fix includeantruntime
These changes eliminate warnings issued during compilation when
ant has been instructed to show warnings and be verbose.

The JCIP (Java Concurrency In Practice) annotations are referred to
by the http client libraries, but are not bundled in them. This tiny
jar eliminates warnings about the use of unknown annotations and enables
IDEs to see and process these annotations.

includeAntRuntime should be set to false on javac tasks so that the
ant runtime isn't on the classpath. This ensures that dependencies
aren't missed because they happened to be included in the version
of ant used for compilation.